+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG
 =========
 
+# 2020.XX.XX
+- ???
+
 # 2020.09.18
 - Fix some pins for dev requirements
 - Add 3.9-dev to actions